feat: migrate benchmarks to bitest
