+++ .github/workflows/benchmark.yaml
@@ -7,6 +7,9 @@ on:
   pull_request:
   workflow_dispatch:
 
+env:
+  CODSPEED: true
+
 jobs:
   benchmarks:
     name: Run benchmarks
