+++ packages/benchmark-wasm/package.json
@@ -6,18 +6,18 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "benchmark": "tsx index.ts",
-    "benchmark-codspeed": "tsx codspeed.ts"
+    "benchmark": "pnpm exec vitest bench"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@codspeed/tinybench-plugin": "^3.1.1",
     "csslsrs": "workspace:*",
-    "tinybench": "^3.0.3",
-    "tsx": "^4.19.1",
     "vscode-css-languageservice": "^6.3.1",
     "vscode-languageserver-textdocument": "^1.0.12"
+  },
+  "dependencies": {
+    "@codspeed/vitest-plugin": "^3.1.1",
+    "vitest": "^2.1.8"
   }
 }
